+++ src/components/modality-options/modality-options.jsx
@@ -1,18 +1,12 @@
 import "./modality-options.css"
 
-const imageModels = [
-    { label: "ResNet", image: "https://production-media.paperswithcode.com/thumbnails/method/6f12f862-c1a9-4278-8ab9-82c658036935.jpg", numPapers: "2157", href: "" },
-    { label: "Visual Transformer", image: "https://production-media.paperswithcode.com/thumbnails/method/1ec5160c-aabe-44b6-a463-3db231949c27.jpg", numPapers: "1430", href: "" },
-    { label: "VGG", image: "https://production-media.paperswithcode.com/thumbnails/method/method-0000000312-bb51f64c.jpg", numPapers: "488", href: ""},
-    { label: "DenseNet", image: "https://production-media.paperswithcode.com/thumbnails/method/method-0000000434-b414a7f9_Ctryexj.jpg", numPapers: "401", href: ""},
-    { label: "MobileNetV2", image: "https://production-media.paperswithcode.com/thumbnails/method/method-0000000170-4df6c581_jjvLsL7.jpg", numPapers: "328", href: "" }
-  ]
 
-const ModalityOptions = () => {
+
+const ModalityOptions = ({imageModels}) => {
   
   return (
 
-    <div className = "container">
+    
 
         <div className = "card-deck card-break">
 
@@ -40,7 +34,7 @@ const ModalityOptions = () => {
 
                         <div className = "card-body">
                             
-                            <div className = "card-subtitle"> { model.numPapers + " papers" }</div>
+                            <div className = "card-subtitle"> { model.numPapers }</div>
                             
                         </div>
                     
@@ -52,7 +46,7 @@ const ModalityOptions = () => {
 
         </div>
     
-    </div>
+    
   )
 }
