+++ src/main.jsx
@@ -1,9 +1,11 @@
 import App from './App.jsx'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
+import ModalityOptions from './components/modality-options/modality-options.jsx'
+import ModalitySection from './components/modality-options/ModalitySection.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
+    <ModalitySection/>
   </React.StrictMode>
 )
